rename and move tests
