+++ src/test/scala/cognite/spark/v1/fdm/utils/FDMViewPropertyTypes.scala
@@ -1,4 +1,4 @@
-package cognite.spark.v1.utils.fdm
+package cognite.spark.v1.fdm.utils
 
 import com.cognite.sdk.scala.v1.fdm.common.properties.PropertyDefinition.ViewCorePropertyDefinition
 import com.cognite.sdk.scala.v1.fdm.common.properties.PropertyType.{DirectNodeRelationProperty, EnumValueMetadata}
